+++ src/main/java/com/gcu/baima/controller/ArticleController.java
@@ -2,9 +2,12 @@ package com.gcu.baima.controller;
 
 
 import cn.hutool.core.lang.UUID;
+import com.gcu.baima.entity.AdmissionPlan;
 import com.gcu.baima.entity.Article;
 import com.gcu.baima.entity.VO.ArticleVo;
+import com.gcu.baima.exception.BaimaException;
 import com.gcu.baima.service.Back.ArticleService;
+import com.gcu.baima.utils.CheckDBUtil;
 import com.gcu.baima.utils.R;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -34,6 +37,8 @@ public class ArticleController {
     @ApiOperation("根据id查询文章,返回文章vo类")
     @GetMapping("{id}")
     public R getById(@PathVariable String id) {
+        //        id不存在
+        if (CheckDBUtil.checkIdEqual(Article.class, id)) throw new BaimaException(201, "id对应的数据不存在");
         ArticleVo articleVo = articleService.getArticleById(id);
         return R.ok().data("article", articleVo);
     }
@@ -41,6 +46,8 @@ public class ArticleController {
     @ApiOperation("添加文章")
     @PostMapping("")
     public R add(@RequestBody Article article) {
+        //        id不存在
+        if (CheckDBUtil.checkIdEqual(Article.class, article.getId())) throw new BaimaException(201, "id对应的数据不存在");
         String id = UUID.randomUUID().toString(true).substring(0, 19);
         article.setId(id);
         article.setPublicTime(new Date());
@@ -57,7 +64,9 @@ public class ArticleController {
 
     @ApiOperation("根据id删除文章")
     @DeleteMapping("{id}")
-    public R removeByid(@PathVariable String id){
+    public R removeByid(@PathVariable String id) {
+        //        id不存在
+        if (CheckDBUtil.checkIdEqual(Article.class, id)) throw new BaimaException(201, "id对应的数据不存在");
         articleService.removeById(id);
         return R.ok();
     }
