+++ src/main/java/com/gcu/baima/controller/ManagerController.java
@@ -3,9 +3,12 @@ package com.gcu.baima.controller;
 
 import cn.hutool.core.bean.BeanUtil;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.gcu.baima.entity.Course;
 import com.gcu.baima.entity.Manager;
 import com.gcu.baima.entity.VO.ManagerVo;
+import com.gcu.baima.exception.BaimaException;
 import com.gcu.baima.service.Back.ManagerService;
+import com.gcu.baima.utils.CheckDBUtil;
 import com.gcu.baima.utils.R;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -33,6 +36,8 @@ public class ManagerController {
     @ApiOperation("根据工作人员id获取工作人员信息")
     @GetMapping("{id}")
     public R getById(@ApiParam("工作人员id") @PathVariable String id) {
+        //        id不存在
+        if (CheckDBUtil.checkIdEqual(Manager.class, id)) throw new BaimaException(201, "id对应的数据不存在");
         Manager manager = managerService.getById(id);
         ManagerVo managerVo = BeanUtil.copyProperties(manager, ManagerVo.class);
         return R.ok().data("manager", managerVo);
@@ -41,6 +46,8 @@ public class ManagerController {
     @ApiOperation("更新信息")
     @PutMapping("")
     public R updateCustom(@ApiParam("修改后的工作人员信息，用工作人员id不能为空") @RequestBody Manager manager) {
+        //        id不存在
+        if (CheckDBUtil.checkIdEqual(Manager.class, manager.getId())) throw new BaimaException(201, "id对应的数据不存在");
         managerService.updateById(manager);
         return R.ok();
     }
@@ -48,6 +55,8 @@ public class ManagerController {
     @ApiOperation("根据用户id删除用户")
     @DeleteMapping("/{id}")
     public R deleteById(@ApiParam("工作人员") @PathVariable String id) {
+        //        id不存在
+        if (CheckDBUtil.checkIdEqual(Manager.class, id)) throw new BaimaException(201, "id对应的数据不存在");
         managerService.removeById(id);
         return R.ok();
     }
