+++ src/main/java/com/gcu/baima/controller/TrialLessonCommentController.java
@@ -5,7 +5,9 @@ import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.gcu.baima.entity.TrialLessonComment;
 import com.gcu.baima.entity.VO.TrialLessonCommentVo;
+import com.gcu.baima.exception.BaimaException;
 import com.gcu.baima.service.Back.TrialLessonCommentService;
+import com.gcu.baima.utils.CheckDBUtil;
 import com.gcu.baima.utils.R;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -32,6 +34,7 @@ public class TrialLessonCommentController {
 
     @DeleteMapping("{id}")
     public R deleteById(@PathVariable String id) {
+        if (!CheckDBUtil.checkIdEqual(TrialLessonComment.class, id)) throw new BaimaException(201, "id对应的数据不存在");
         commentService.removeById(id);
         return R.ok();
     }
@@ -53,8 +56,8 @@ public class TrialLessonCommentController {
     //    判断是否已评价该门课程, 试听课程ID、用户ID
     @ApiOperation("判断是否已评价该门课程")
     @GetMapping("isRate")
-    public R isRate(String trialLesson, String customerId) {
-        Boolean rate = commentService.isRate(trialLesson, customerId);
+    public R isRate(String courseId, String customerId) {
+        Boolean rate = commentService.isRate(courseId, customerId);
         if (rate) return R.ok().message("已评价").data("isRate", 1);
         return R.ok().message("未评价").data("isRate", 0);
     }
